V.0.5.4.0.0_Потребности_2: полное сохранение проекта; актуализация версий до 0.5.4.0.0 (version.txt, pyproject, frontend/constants, README). Mypy-доводка 584->545: 39 мелочей закрыты без вмешательства в логику (Optional-дефолты, -> None, no-any-return через bool()/float()/cast(), var-annotated словари, class-level аттрибуты singleton, аннотации choice/choices). Гейты: ruff по всем 15 затронутым файлам OK, импорты ядра OK, 179 pytest passed (combat/content_policy/action_commitment/decision_calibration/commitment_ssm/speech_exposure micro). Долг сессии: артефакты Р-Г GC-DIALOGUE-01 (npc_dialogue_subscriber радиус-журнал, test_self_talk_sentinel регрессия), логи аудита/сцен/world_tick, diagnostics-пробы, ADR-O-384/385 impact, ТЗ mvp_secret_tavern.

+++ backend/app/services/npc/decision/profile_math.py
@@ -37,10 +37,10 @@ def _shape_drive(drive: float) -> float:
     # Нелинейность: power curve с центром на NEUTRAL_DRIVE
     if d > NEUTRAL_DRIVE:
         excess = (d - NEUTRAL_DRIVE) / (1.0 - NEUTRAL_DRIVE)
-        return NEUTRAL_DRIVE + (1.0 - NEUTRAL_DRIVE) * (excess ** (1.0 / POWER_ALPHA))
+        return float(NEUTRAL_DRIVE + (1.0 - NEUTRAL_DRIVE) * (excess ** (1.0 / POWER_ALPHA)))
     elif d < NEUTRAL_DRIVE:
         deficit = (NEUTRAL_DRIVE - d) / NEUTRAL_DRIVE
-        return NEUTRAL_DRIVE * (1.0 - deficit**POWER_ALPHA)
+        return float(NEUTRAL_DRIVE * (1.0 - deficit**POWER_ALPHA))
     else:
         return NEUTRAL_DRIVE
